ingest/Minio/python-etl/requirements.txt
# Python Dependencies

psycopg2-binary
minio
pandas
pyarrow
sqlalchemy
python-dotenv